chore(api): replace ts-jest with swc
