+++ services/api/package.json
@@ -76,6 +76,9 @@
     "@prisma/generator-helper": "3.15.2",
     "@prisma/migrate": "3.15.2",
     "@prisma/sdk": "3.15.2",
+    "@side/jest-runtime": "^1.1.0",
+    "@swc/core": "^1.3.39",
+    "@swc/jest": "^0.2.24",
     "@types/bcrypt": "5.0.0",
     "@types/cookie-parser": "1.4.3",
     "@types/faker": "5.5.9",
