+++ services/api/src/prisma/prisma-module-config.interface.ts
@@ -1,0 +1,15 @@
+import { FactoryProvider, ModuleMetadata } from "@nestjs/common";
+
+export interface PrismaModuleConfig {
+  databaseUrl: string;
+}
+
+export interface PrismaAsyncModuleConfig extends Pick<ModuleMetadata, "imports"> {
+  useFactory: (...args: any[]) => PrismaModuleConfig | Promise<PrismaModuleConfig>;
+  inject?: FactoryProvider["inject"];
+}
+
+export interface PrismaAsyncModuleConfig extends Pick<ModuleMetadata, "imports"> {
+  useFactory: (...args: any[]) => PrismaModuleConfig | Promise<PrismaModuleConfig>;
+  inject?: FactoryProvider["inject"];
+}
